+++ spec/spec_helper.rb
@@ -17,95 +17,6 @@
 #
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 
-require 'capybara/rspec'
-require 'capybara-screenshot/rspec'
-require 'capybara/email/rspec'
-require 'database_cleaner'
-require 'webmock/rspec'
-require 'shoulda-matchers'
-require 'devise'
-require 'factory_bot'
-require 'axe/rspec'
-
-require 'selenium/webdriver'
-Capybara.javascript_driver = :headless_chrome
-Capybara.ignore_hidden_elements = false
-
-Capybara.register_driver :chrome do |app|
-  Capybara::Selenium::Driver.new(app, browser: :chrome)
-end
-
-Capybara.register_driver :headless_chrome do |app|
-  options = Selenium::WebDriver::Chrome::Options.new
-  options.add_argument('--headless')
-  options.add_argument('--window-size=1440,900')
-
-  capabilities = Selenium::WebDriver::Remote::Capabilities.chrome(
-    chromeOptions: { args: ['disable-dev-shm-usage', 'disable-software-rasterizer', 'mute-audio', 'window-size=1440,900'] }
-  )
-
-  download_path = Capybara.save_path
-  # Chromedriver 77 requires setting this for headless mode on linux
-  # Different versions of Chrome/selenium-webdriver require setting differently - just set them all
-  options.add_preference('download.default_directory', download_path)
-  options.add_preference(:download, default_directory: download_path)
-
-  Capybara::Selenium::Driver.new(app,
-    browser: :chrome,
-    desired_capabilities: capabilities,
-    options: options).tap do |driver|
-    # Set download dir for Chrome < 77
-    driver.browser.download_path = download_path
-  end
-end
-
-# FIXME: remove this line when https://github.com/rspec/rspec-rails/issues/1897 has been fixed
-Capybara.server = :puma, { Silent: true }
-
-Capybara.default_max_wait_time = 2
-
-# Save a snapshot of the HTML page when an integration test fails
-Capybara::Screenshot.autosave_on_failure = true
-# Keep only the screenshots generated from the last failing test suite
-Capybara::Screenshot.prune_strategy = :keep_last_run
-# Tell Capybara::Screenshot how to take screenshots when using the headless_chrome driver
-Capybara::Screenshot.register_driver :headless_chrome do |driver, path|
-  driver.browser.save_screenshot(path)
-end
-
-# Requires supporting ruby files with custom matchers and macros, etc,
-# in spec/support/ and its subdirectories.
-Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each { |f| require f }
-Dir[Rails.root.join('spec', 'factories', '**', '*.rb')].each { |f| require f }
-
-VCR.configure do |c|
-  c.ignore_localhost = true
-  c.hook_into :webmock
-  c.cassette_library_dir = 'spec/fixtures/cassettes'
-  c.configure_rspec_metadata!
-  c.ignore_hosts 'test.host', 'chromedriver.storage.googleapis.com'
-end
-
-DatabaseCleaner.strategy = :transaction
-
-TPS::Application.load_tasks
-Rake.application.options.trace = false
-
-include Warden::Test::Helpers
-
-include SmartListing::Helper
-include SmartListing::Helper::ControllerExtensions
-
-module SmartListing
-  module Helper
-    def view_context
-      'mock'
-    end
-  end
-end
-
-WebMock.disable_net_connect!(allow_localhost: true)
-
 RSpec.configure do |config|
   config.filter_run_excluding disable: true
   config.color = true
@@ -118,26 +29,6 @@ RSpec.configure do |config|
   # Fix the seed not changing between runs when using Spring
   # See https://github.com/rails/spring/issues/113
   config.seed = srand % 0xFFFF unless ARGV.any? { |arg| arg =~ /seed/ || arg =~ /rand:/ }
-
-  config.include FactoryBot::Syntax::Methods
-
-  config.before(:each) do
-    Flipper.enable(:instructeur_bypass_email_login_token)
-  end
-
-  config.before(:all) {
-    Rake.verbose false
-
-    Warden.test_mode!
-
-    Typhoeus::Expectation.clear
-
-    ActionMailer::Base.deliveries.clear
-
-    ActiveStorage::Current.host = 'http://test.host'
-
-    Geocoder.configure(lookup: :test)
-  }
 
   RSpec::Matchers.define :have_same_attributes_as do |expected, options|
     match do |actual|
