+++ spec/support/smart_listing.rb
@@ -1,0 +1,10 @@
+include SmartListing::Helper
+include SmartListing::Helper::ControllerExtensions
+
+module SmartListing
+  module Helper
+    def view_context
+      'mock'
+    end
+  end
+end
